Added login and logout to correct components
