+++ src/Components/ProfileBox/ProfileBox.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import profileImage from '../../images/template.png';
+import { GoogleLogout } from 'react-google-login';
 import {
   StyledMainWrapper,
   StyledFlexWrapper,
@@ -10,6 +11,16 @@ import {
 } from './StyledProfileBox'
 
 const ProfileBox = () => {
+
+  const signOut = () => {
+    const auth2 = window.gapi.auth2.getAuthInstance();
+    auth2.signOut().then(() => {
+      console.log('You have been logged out!');
+    });
+    window.location = "/login"
+  }
+
+
   return (
     <StyledMainWrapper>
       <StyledFlexWrapper>
@@ -20,6 +31,11 @@ const ProfileBox = () => {
         <StyledButton>
           Min Profil
         </StyledButton>
+        <GoogleLogout
+          buttonText="Logout"
+          onLogoutSuccess={signOut}
+        >
+        </GoogleLogout>
       </StyledProfileLinksContainer>
     </StyledMainWrapper>
   )
